Move `unified` plugin into a separate file
